Make CRAS structure sizes architecture-independent.
This allows x86 clients to connect to x86-64 server.

Get rid of size_t, long, and timespec, whose size depend on the architecture.
The only assumption we make is sizeof(int) = 4.

Volume is now an unsigned int (was size_t).
Input gain is now an int (was long).
We replace uses of timespec by a fixed-size, 16 bytes long cras_timespec.